requirements.txt
langchain==0.3.18
langchain-openai==0.3.5
langchain-anthropic==0.3.7
langgraph==0.3.21
langmem==0.0.19
python-dotenv==1.0.1
mem0ai
pyautogen
vecs
supabase